Add minor bugfix improvements
Fix for disc reading turned out to be in a newly found bug in libiosuhax, or actually it's more of a Wii U quirk that was just hidden up till now.

Further more, it now more properly restores the state of the application after disc dumping and might also fix a potential issue with saves stored on HDDs.

Also return to previous buffer size since I'm pretty sure it causes dumping to randomly fail.